+++ examples/kyb-app/src/common/utils/query-client.ts
@@ -1,3 +1,10 @@
 import { QueryClient } from '@tanstack/react-query';
 
-export const queryClient = new QueryClient({});
+export const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      retry: false,
+      retryOnMount: false,
+    },
+  },
+});